bsssss ich bin ein fleißiges Bienchen

+++ src/FilesystemStorageSystem/MessageSubscriptionStorage.cs
@@ -4,16 +4,29 @@ using StorageSystem;
 
 namespace FilesystemStorageSystem;
 
-public class MessageSubscriptionStorage(Storage storage) : IMessageSubscriptionStorage
+public class MessageSubscriptionStorage : IMessageSubscriptionStorage
 {
-    private readonly string _messageSubscriptionsPath = storage.GetBasePath("FileStorage/MessageSubscriptions");
-    
+    private readonly string _messageSubscriptionsPath;
+    private readonly JsonSerializerSettings _newtonSoftDefaultSettings;
+
+    public MessageSubscriptionStorage(Storage storage)
+    {
+        _messageSubscriptionsPath = storage.GetBasePath("FileStorage/MessageSubscriptions");
+        
+        _newtonSoftDefaultSettings = new JsonSerializerSettings
+        {
+            TypeNameHandling = TypeNameHandling.Auto,
+            TypeNameAssemblyFormatHandling = TypeNameAssemblyFormatHandling.Simple,
+            Formatting = Formatting.Indented,
+        };
+    }
+
     public Task<IEnumerable<MessageSubscription>> GetAllMessageSubscriptions()
     {
-        return Task.FromResult(Directory.GetFiles(_messageSubscriptionsPath, "*.json").Select(file =>
+        return Task.FromResult(Directory.GetFiles(_messageSubscriptionsPath, "message_*.json").Select(file =>
         {
             var content = File.ReadAllText(file);
-            var messageSubscription = JsonConvert.DeserializeObject<MessageSubscription>(content)!;
+            var messageSubscription = JsonConvert.DeserializeObject<MessageSubscription>(content, _newtonSoftDefaultSettings)!;
             return messageSubscription;
         }));
     }
@@ -40,7 +53,7 @@ public class MessageSubscriptionStorage(Storage storage) : IMessageSubscriptionS
     {
         var randomIdOrInstanceId = messageSubscription.ProcessInstanceId ?? Guid.NewGuid();
         var fullFileName = Path.Combine(_messageSubscriptionsPath, $"message_{messageSubscription.RelatedDefinitionId}_{randomIdOrInstanceId}.json");
-        var data = JsonConvert.SerializeObject(messageSubscription, Formatting.Indented);
+        var data = JsonConvert.SerializeObject(messageSubscription, _newtonSoftDefaultSettings);
         return File.WriteAllTextAsync(fullFileName, data);
     }
 
@@ -56,23 +69,27 @@ public class MessageSubscriptionStorage(Storage storage) : IMessageSubscriptionS
     }
 
 
-    public Task RemoveProcessMessageSubscriptions(string relatedDefinitionId)
+    public Task RemoveAllProcessMessageSubscriptionsWithNoInstancedId(string metaDefinitionId)
     {
-        var files = Directory.GetFiles(_messageSubscriptionsPath, $"message_{relatedDefinitionId}_*.json");
+        var files = Directory.GetFiles(_messageSubscriptionsPath, $"message_{metaDefinitionId}_*.json");
         foreach (var file in files)
         {
-            File.Delete(file);
+            var subscription = JsonConvert.DeserializeObject<MessageSubscription>(File.ReadAllText(file),_newtonSoftDefaultSettings);
+            if (subscription?.ProcessInstanceId == null || subscription.ProcessInstanceId == Guid.Empty)
+                File.Delete(file);
         }
 
         return Task.CompletedTask;
     }
 
-    public Task RemoveProcessSignalSubscriptions(string relatedDefinitionId)
+    public Task RemoveAllProcessSignalSubscriptionsWithNoInstanceId(string relatedDefinitionId)
     {
         var files = Directory.GetFiles(_messageSubscriptionsPath, $"signal_{relatedDefinitionId}_*.json");
         foreach (var file in files)
         {
-            File.Delete(file);
+            var subscription = JsonConvert.DeserializeObject<SignalSubscription>(File.ReadAllText(file),_newtonSoftDefaultSettings);
+            if (subscription?.ProcessInstanceId == null || subscription.ProcessInstanceId == Guid.Empty)
+                File.Delete(file);
         }
         return Task.CompletedTask;
     }
@@ -80,7 +97,7 @@ public class MessageSubscriptionStorage(Storage storage) : IMessageSubscriptionS
     public void AddSignalSubscription(SignalSubscription signalSubscription)
     {
         var fullFileName = Path.Combine(_messageSubscriptionsPath, $"signal_{signalSubscription.RelatedDefinitionId}_{Guid.NewGuid()}.json");
-        var data = JsonConvert.SerializeObject(signalSubscription, Formatting.Indented);
+        var data = JsonConvert.SerializeObject(signalSubscription, _newtonSoftDefaultSettings);
         File.WriteAllText(fullFileName, data);
     }
 
@@ -91,5 +108,61 @@ public class MessageSubscriptionStorage(Storage storage) : IMessageSubscriptionS
         {
             File.Delete(file);
         }
+    }
+
+    public async Task<IEnumerable<UserTaskSubscription>> GetAllUserTasks(Guid userId)
+    {
+        var files = Directory.GetFiles(_messageSubscriptionsPath, $"usertask_*.json");
+        var ret = new List<UserTaskSubscription>();
+        foreach (var file in files)
+        {
+            var content = File.ReadAllText(file);
+            var userTaskSubscription = JsonConvert.DeserializeObject<UserTaskSubscription>(content,_newtonSoftDefaultSettings)!;
+            ret.Add(userTaskSubscription);
+        }
+
+        return ret;
+    }
+
+    public Task AddUserTaskSubscription(UserTaskSubscription userTasks)
+    {
+        var fullFileName = Path.Combine(_messageSubscriptionsPath, $"usertask_{userTasks.Id}.json");
+        var data = JsonConvert.SerializeObject(userTasks, _newtonSoftDefaultSettings);
+        return File.WriteAllTextAsync(fullFileName, data);
+    }
+
+    public Task RemoveUserTaskSubscription(Guid userTaskSubscriptionId)
+    {
+        var files = Directory.GetFiles(_messageSubscriptionsPath, $"usertask_{userTaskSubscriptionId}.json");
+        foreach (var file in files)
+        {
+            File.Delete(file);
+        }
+
+        return Task.CompletedTask;
+    }
+
+    public void RemoveAllUserTaskSubscriptionsByInstanceId(Guid instanceId)
+    {
+        var files = Directory.GetFiles(_messageSubscriptionsPath, $"usertask_*.json");
+        foreach (var file in files)
+        {
+            var subscription = JsonConvert.DeserializeObject<UserTaskSubscription>(File.ReadAllText(file),_newtonSoftDefaultSettings)!;
+            if (subscription.ProcessInstanceId == instanceId)
+                File.Delete(file);
+        }
+    }
+
+    public Task RemoveAllUserTaskSubscriptionsWithNoInstanceId(string relatedDefinitionId)
+    {
+        var files = Directory.GetFiles(_messageSubscriptionsPath, $"usertask_{relatedDefinitionId}_*.json");
+        foreach (var file in files)
+        {
+            var subscription = JsonConvert.DeserializeObject<UserTaskSubscription>(File.ReadAllText(file),_newtonSoftDefaultSettings);
+            if (subscription?.ProcessInstanceId == null || subscription.ProcessInstanceId == Guid.Empty)
+                File.Delete(file);
+        }
+
+        return Task.CompletedTask;
     }
 }
